+++ src/PackageReferenceVersionToAttribute/ProjectConverter.cs
@@ -11,7 +11,6 @@ namespace PackageReferenceVersionToAttribute
     using System.Threading.Tasks;
     using System.Xml;
     using System.Xml.Linq;
-    using System.Xml.XPath;
     using Microsoft.Extensions.Logging;
     using Microsoft.Extensions.Options;