+++ src/PackageReferenceVersionToAttributeTool/ProgramCommand.cs
@@ -4,15 +4,8 @@
 
 namespace PackageReferenceVersionToAttributeTool
 {
-    using System;
-    using System.Collections.Generic;
     using System.CommandLine;
     using System.CommandLine.NamingConventionBinder;
-    using System.Reflection;
-    using System.Threading.Tasks;
-    using Microsoft.Extensions.DependencyInjection;
-    using Microsoft.Extensions.Logging;
-    using PackageReferenceVersionToAttribute;
 
     /// <summary>
     /// Program command.
@@ -51,82 +44,13 @@ namespace PackageReferenceVersionToAttributeTool
             this.Add(forceOption);
             this.Add(dryRunOption);
 
-            // Add validation
-            this.AddValidator(result =>
-            {
-                bool backup = result.GetValueForOption(backupOption);
-                bool force = result.GetValueForOption(forceOption);
-                bool dryRun = result.GetValueForOption(dryRunOption);
-
-                var options = new ProjectConverterOptions
-                {
-                    Backup = backup,
-                    Force = force,
-                    DryRun = dryRun,
-                };
-
-                var validator = new ProjectConverterOptionsValidator();
-                var validationResult = validator.Validate(nameof(ProjectConverterOptions), options);
-                if (validationResult.Failed)
-                {
-                    result.ErrorMessage = validationResult.FailureMessage;
-                }
-            });
+            // validate
+            var commandValidator = new ProgramCommandLineOptionsValidator(backupOption, forceOption, dryRunOption);
+            this.AddValidator(commandValidator.Validate);
 
             // Set the handler for the command
-            this.Handler = CommandHandler.Create<ProgramCommandLineOptions>(async (options) =>
-            {
-                if (options.Version)
-                {
-                    var version = Assembly.GetExecutingAssembly().GetName().Version;
-                    Console.WriteLine($"Version: {version}");
-                    return;
-                }
-
-                foreach (var input in options.Inputs)
-                {
-                    await ConvertPackageReferencesAsync(input, options);
-                }
-            });
-        }
-
-        private static async Task ConvertPackageReferencesAsync(
-            string input, ProjectConverterOptions options)
-        {
-            FilePatternMatcher filePatternMatcher = new();
-            List<string> matchingFiles = filePatternMatcher.GetMatchingFiles(input);
-            if (matchingFiles.Count == 0)
-            {
-                Console.WriteLine($"No matching files found for pattern: {input}");
-                return;
-            }
-
-            if (options.Backup)
-            {
-                Console.WriteLine("Backup option is enabled.");
-            }
-
-            if (options.Force)
-            {
-                Console.WriteLine("Force option is enabled.");
-            }
-
-            if (options.DryRun)
-            {
-                Console.WriteLine("Dry run mode is enabled. No changes will be made.");
-            }
-
-            using var serviceProvider = new ServiceCollection()
-                .AddSingleton(Microsoft.Extensions.Options.Options.Create(options))
-                .AddLogging(configure => configure.AddConsole())
-                .AddSingleton<ProjectConverter>()
-                .AddSingleton<IFileService, FileService>()
-                .AddSingleton<ISourceControlService, NullSourceControlService>()
-                .BuildServiceProvider();
-
-            var projectConverter = serviceProvider.GetRequiredService<ProjectConverter>();
-
-            await projectConverter.ConvertAsync(matchingFiles);
+            this.Handler = CommandHandler.Create<ProgramCommandLineOptions>(
+                ProgramCommandHandler.HandleAsync);
         }
     }
 }