+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## TBD
+
+- Added support for converting projects within sln files on the command line
+
 ## v1.0.1104.33 (November 4<sup>th</sup>, 2024)
 
 - Added input validation for the mutually exclusive --backup and --dry-run command line options